test power_mean

+++ engine/src/agents/config/searchsettings.h
@@ -89,7 +89,7 @@ struct SearchSettings
     uint32_t maxAtVisit;
     uint32_t switchingMaxOperatorAtNode;
     float minimaxWeight;
-    float power_mean;
+    double power_mean;
     SearchSettings();
 
 };
